move qcbor test functions and add test of array issue
