+++ test/test_MockBPA_Codecs.c
@@ -258,72 +258,6 @@ void test_bsl_mock_encode_bundle(void)
     MockBPA_Bundle_Deinit(&bundle);
 }
 
-void test_qcbor_decode_without_head(void)
-{
-    string_t in_text;
-    string_init_set_str(in_text, "58"); // not a full head
-    BSL_Data_t in_data;
-    BSL_Data_Init(&in_data);
-    TEST_ASSERT_EQUAL_INT_MESSAGE(0, BSL_TestUtils_DecodeBase16(&in_data, in_text),
-                                  "BSL_TestUtils_DecodeBase16() failed");
-
-    QCBORDecodeContext decoder;
-    QCBORDecode_Init(&decoder, (UsefulBufC) { in_data.ptr, in_data.len }, QCBOR_DECODE_MODE_NORMAL);
-    QCBORItem item;
-    TEST_ASSERT_EQUAL_INT_MESSAGE(QCBOR_ERR_HIT_END, QCBORDecode_GetNext(&decoder, &item),
-                                  "QCBORDecode_VGetNext() failed");
-    TEST_ASSERT_EQUAL_INT(QCBOR_TYPE_NONE, item.uDataType);
-    TEST_ASSERT_EQUAL_INT(1, QCBORDecode_Tell(&decoder));
-    TEST_ASSERT_EQUAL_INT(QCBOR_SUCCESS, QCBORDecode_Finish(&decoder));
-
-    BSL_Data_Deinit(&in_data);
-    string_clear(in_text);
-}
-
-void test_qcbor_decode_only_head(void)
-{
-    string_t in_text;
-    string_init_set_str(in_text, "586C"); // just a full head
-    BSL_Data_t in_data;
-    BSL_Data_Init(&in_data);
-    TEST_ASSERT_EQUAL_INT_MESSAGE(0, BSL_TestUtils_DecodeBase16(&in_data, in_text),
-                                  "BSL_TestUtils_DecodeBase16() failed");
-
-    QCBORDecodeContext decoder;
-    QCBORDecode_Init(&decoder, (UsefulBufC) { in_data.ptr, in_data.len }, QCBOR_DECODE_MODE_NORMAL);
-    QCBORItem item;
-    TEST_ASSERT_EQUAL_INT_MESSAGE(QCBOR_ERR_HIT_END, QCBORDecode_GetNext(&decoder, &item),
-                                  "QCBORDecode_VGetNext() failed");
-    TEST_ASSERT_EQUAL_INT(QCBOR_TYPE_NONE, item.uDataType);
-    TEST_ASSERT_EQUAL_INT(2, QCBORDecode_Tell(&decoder));
-    TEST_ASSERT_EQUAL_INT(QCBOR_SUCCESS, QCBORDecode_Finish(&decoder));
-
-    BSL_Data_Deinit(&in_data);
-    string_clear(in_text);
-}
-
-void test_qcbor_decode_with_head(void)
-{
-    string_t in_text;
-    string_init_set_str(in_text, "586C616263646566"); // front of a bstr value
-    BSL_Data_t in_data;
-    BSL_Data_Init(&in_data);
-    TEST_ASSERT_EQUAL_INT_MESSAGE(0, BSL_TestUtils_DecodeBase16(&in_data, in_text),
-                                  "BSL_TestUtils_DecodeBase16() failed");
-
-    QCBORDecodeContext decoder;
-    QCBORDecode_Init(&decoder, (UsefulBufC) { in_data.ptr, in_data.len }, QCBOR_DECODE_MODE_NORMAL);
-    QCBORItem item;
-    TEST_ASSERT_EQUAL_INT_MESSAGE(QCBOR_ERR_HIT_END, QCBORDecode_GetNext(&decoder, &item),
-                                  "QCBORDecode_VGetNext() failed");
-    TEST_ASSERT_EQUAL_INT(QCBOR_TYPE_NONE, item.uDataType);
-    TEST_ASSERT_EQUAL_INT(2, QCBORDecode_Tell(&decoder));
-    TEST_ASSERT_EQUAL_INT(QCBOR_ERR_EXTRA_BYTES, QCBORDecode_Finish(&decoder));
-
-    BSL_Data_Deinit(&in_data);
-    string_clear(in_text);
-}
-
 TEST_CASE("8202820102")                   // IPN scheme
 TEST_CASE("821A00010000A203426869041819") // unknown scheme with complex data
 void test_bsl_loopback_eid(const char *hexdata)
